lib/nlattr.c: Fix stringop overflow warning in nla_strcmp()
strlen() returns a size_t value, so store it in a size_t variable.

+++ lib/nlattr.c
@@ -303,7 +303,7 @@ int nla_memcmp(const struct nlattr *nla, const void *data,
  */
 int nla_strcmp(const struct nlattr *nla, const char *str)
 {
-	int len = strlen(str);
+	size_t len = strlen(str);
 	char *buf = nla_data(nla);
 	int attrlen = nla_len(nla);
 	int d;
